requirements.txt
django==3.2
pytest==6.2.4
pytest-mock==3.6.1
pytest-django==4.4.0
djangorestframework==3.12.4

